pyproject.toml
[tool.poetry]
name = "djangify"
version = "0.1.0"
description = "A Django and Next.js portfolio and blog application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
django = "^5.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
python-magic = "^0.4.27"
django-summernote = "^0.8.20.0"
pillow = "^10.2.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
django-filter = "^23.5"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
django-debug-toolbar = "^4.4.6"
drf-spectacular = "^0.27.2"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest-django = "^4.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
